django-generator-pdf-qrcode-main/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
colorama==0.4.4
Django==3.2.7
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-pdfkit==0.3.1
django-phonenumber-field==5.2.0
pdfkit==0.6.0
phonenumbers==8.12.33
Pillow==8.3.2
pytz==2021.1
qrcode==7.3
soupsieve==2.2.1
sqlparse==0.4.2
wkhtmltopdf==0.2
